SMGCPA: Accelerate constructing of a new SMG by copying same hasValueEdges at SMGJoinSubSMGs. Simplify getNullEdgesMapOffsetToSizeForObject() because SMGHasValueEdges already merges neighboring zero value fields
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33924 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/smg/graphs/UnmodifiableSMG.java
@@ -61,7 +61,7 @@ public interface UnmodifiableSMG {
 
   int getSizeofPtrInBits();
 
-  TreeMap<Long, Integer> getNullEdgesMapOffsetToSizeForObject(SMGObject pObj);
+  TreeMap<Long, Long> getNullEdgesMapOffsetToSizeForObject(SMGObject pObj);
 
   boolean isPointer(SMGValue value);
